(PC-41342)[PRO] refactor: delete isSwitchVenueEnabled and old Sitemap
