--- pro/src/pages/Sitemap/NewSitemap.module.scss
@@ -1,38 +1,0 @@
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_size.scss" as size;
-
-.sitemap-list {
-  margin-top: var(--size-spacing-xxl);
-  list-style-type: disc;
-  list-style-position: inside;
-  width: 100%;
-  margin-bottom: var(--size-spacing-xxl);
-
-  &-item {
-    margin-bottom: var(--size-spacing-s);
-  }
-
-  &-subtitle {
-    @include fonts.body;
-  }
-}
-
-.sitemap-sub-list {
-  list-style-type: disc;
-  list-style-position: inside;
-  padding-left: var(--size-spacing-l);
-  margin-top: var(--size-spacing-s);
-}
-
-.sitemap-link {
-  @include fonts.body;
-
-  text-decoration: underline;
-}
-
-.sitemap-section-link {
-  @include fonts.title4;
-
-  text-decoration: underline;
-}
