+++ pro/src/pages/Sitemap/Sitemap.tsx
@@ -2,6 +2,7 @@ import { Link, useNavigate } from 'react-router'
 
 import { BasicLayout } from '@/app/App/layouts/BasicLayout/BasicLayout'
 import { useAppSelector } from '@/commons/hooks/useAppSelector'
+import { selectSelectedPartnerPageId } from '@/commons/store/nav/selector'
 import { selectCurrentOffererId } from '@/commons/store/offerer/selectors'
 import { Button } from '@/design-system/Button/Button'
 import {
@@ -13,9 +14,163 @@ import fullBackIcon from '@/icons/full-back.svg'
 
 import styles from './Sitemap.module.scss'
 
+interface SitemapLink {
+  title: string
+  path: string | null
+  children: SitemapLink[]
+  hidden?: boolean
+}
+
 export const Sitemap = () => {
   const selectedOffererId = useAppSelector(selectCurrentOffererId)
   const navigate = useNavigate()
+  const reduxStoredPartnerPageId = useAppSelector(selectSelectedPartnerPageId)
+  const selectedPartnerVenue = useAppSelector(
+    (state) => state.user.selectedPartnerVenue
+  )
+
+  const sitemapLinks: SitemapLink[] = [
+    {
+      title: 'Hub',
+      path: '/hub',
+      children: [],
+    },
+    {
+      title: 'Accueil',
+      path: '/accueil',
+      children: [
+        {
+          title: 'Créer une offre',
+          path: `/offre/individuelle/creation/description`,
+          children: [],
+        },
+        {
+          title: 'Individuel',
+          path: null,
+          children: [
+            { title: 'Offres', path: '/offres', children: [] },
+            { title: 'Réservations', path: '/reservations', children: [] },
+            { title: 'Guichet', path: '/guichet', children: [] },
+            {
+              title: "Page sur l'application (si offre individuelle créée)",
+              path: `/structures/${selectedOffererId}/lieux/${reduxStoredPartnerPageId}/page-partenaire`,
+              hidden: !selectedPartnerVenue?.hasPartnerPage,
+              children: [],
+            },
+          ],
+        },
+        {
+          title: 'Collectif',
+          path: null,
+          children: [
+            {
+              title: 'Offres vitrines',
+              path: '/offres/vitrines',
+              children: [],
+            },
+            {
+              title: 'Offres réservables',
+              path: '/offres/collectives',
+              children: [],
+            },
+            {
+              title: 'Page sur ADAGE (si validé ADAGE)',
+              path: `/structures/${selectedOffererId}/lieux/${selectedPartnerVenue?.id}/collectif`,
+              hidden: !selectedPartnerVenue?.allowedOnAdage,
+              children: [],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      title: 'Espace administration',
+      path: '/remboursements',
+      children: [
+        {
+          title: 'Gestion financière',
+          path: null,
+          children: [
+            { title: 'Justificatifs', path: '/remboursements', children: [] },
+            {
+              title: 'Informations bancaires',
+              path: '/remboursements/informations-bancaires',
+              children: [],
+            },
+            {
+              title: 'Chiffre d’affaires',
+              path: '/remboursements/revenus',
+              children: [],
+            },
+          ],
+        },
+        {
+          title: "Données d'activité",
+          path: null,
+          children: [
+            {
+              title: 'Individuel',
+              path: '/donnees-activité/individuel',
+              children: [],
+            },
+            {
+              title: 'Collectif',
+              path: '/donnees-activité/collectif',
+              children: [],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      title: 'Collaborateurs',
+      path: '/collaborateurs',
+      children: [],
+    },
+    {
+      title: 'Paramètres généraux',
+      path: `/structures/${selectedOffererId}/lieux/${selectedPartnerVenue?.id}/parametres`,
+      children: [],
+    },
+    {
+      title: 'Profil',
+      path: '/profil',
+      children: [],
+    },
+  ]
+  const renderSitemapItems = (items: SitemapLink[]) => {
+    return items
+      .filter((item) => !item.hidden)
+      .map((link) => (
+        <li
+          key={link.path ?? link.title}
+          className={styles['sitemap-list-item']}
+        >
+          {link.path ? (
+            <Link
+              to={link.path}
+              className={
+                link.children.length > 0
+                  ? styles['sitemap-section-link']
+                  : styles['sitemap-link']
+              }
+            >
+              {link.title}
+            </Link>
+          ) : (
+            <span className={styles['sitemap-list-subtitle']}>
+              {link.title}
+            </span>
+          )}
+
+          {link.children.length > 0 && (
+            <ul className={styles['sitemap-sub-list']}>
+              {renderSitemapItems(link.children)}
+            </ul>
+          )}
+        </li>
+      ))
+  }
 
   return (
     <BasicLayout mainHeading="Plan du site" isFullPage={true}>
@@ -28,96 +183,7 @@ export const Sitemap = () => {
         label="Retour"
       />
       <ul className={styles['sitemap-list']} data-testid="sitemap">
-        <li className={styles['sitemap-list-item']}>
-          <Link
-            to={`/offre/creation?structure=${selectedOffererId}`}
-            className={styles['sitemap-link']}
-          >
-            Créer une offre
-          </Link>
-        </li>
-        <li className={styles['sitemap-list-item']}>
-          <Link to="/accueil" className={styles['sitemap-link']}>
-            Accueil
-          </Link>
-        </li>
-        <li className={styles['sitemap-list-item']}>
-          <span className={styles['sitemap-list-title']}>Individuel</span>
-          <ul className={styles['sitemap-sub-list']}>
-            <li className={styles['sitemap-list-item']}>
-              <Link to="/offres" className={styles['sitemap-link']}>
-                Offres
-              </Link>
-            </li>
-            <li className={styles['sitemap-list-item']}>
-              <Link to="/reservations" className={styles['sitemap-link']}>
-                Réservations
-              </Link>
-            </li>
-            <li className={styles['sitemap-list-item']}>
-              <Link to="/guichet" className={styles['sitemap-link']}>
-                Guichet
-              </Link>
-            </li>
-          </ul>
-        </li>
-
-        <li className={styles['sitemap-list-item']}>
-          <span className={styles['sitemap-list-title']}>Collectif</span>
-          <ul className={styles['sitemap-sub-list']}>
-            <li className={styles['sitemap-list-item']}>
-              <Link to="/offres/collectives" className={styles['sitemap-link']}>
-                Offres
-              </Link>
-            </li>
-            <li className={styles['sitemap-list-item']}>
-              <Link
-                to="/reservations/collectives"
-                className={styles['sitemap-link']}
-              >
-                Réservations
-              </Link>
-            </li>
-          </ul>
-        </li>
-        <li className={styles['sitemap-list-item']}>
-          <span className={styles['sitemap-list-title']}>
-            Gestion financière
-          </span>
-          <ul className={styles['sitemap-sub-list']}>
-            <li className={styles['sitemap-list-item']}>
-              <Link to="/remboursements" className={styles['sitemap-link']}>
-                Justificatifs
-              </Link>
-            </li>
-            <li className={styles['sitemap-list-item']}>
-              <Link
-                to="/remboursements/informations-bancaires"
-                className={styles['sitemap-link']}
-              >
-                Informations bancaires
-              </Link>
-            </li>
-            <li className={styles['sitemap-list-item']}>
-              <Link
-                to="/remboursements/revenus"
-                className={styles['sitemap-link']}
-              >
-                Chiffre d’affaires
-              </Link>
-            </li>
-          </ul>
-        </li>
-        <li className={styles['sitemap-list-item']}>
-          <Link to="/collaborateurs" className={styles['sitemap-link']}>
-            Collaborateurs
-          </Link>
-        </li>
-        <li className={styles['sitemap-list-item']}>
-          <Link to="/profil" className={styles['sitemap-link']}>
-            Profil
-          </Link>
-        </li>
+        {renderSitemapItems(sitemapLinks)}
       </ul>
     </BasicLayout>
   )
